Fix keys reference link in ARM
There's a statement that says: "Here is an example (for illustration
purposes only) of some possible entries in a [keys] list:", and that
links to the wrong "keys" statement (it links to the TSIG keys section).

Remove the reference, as we are already in the right section.

Note: This commit is empty because 9.16 was not affected by this.

(cherry-picked from commit 07c70ea2bd2b4ba5051a74cd18cc4030628360c8)